[refactor]PMA-39: Implement Auth token interceptor
