+++ src/app/app.module.ts
@@ -6,7 +6,7 @@ import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
-import { httpInterceptorProviders } from './shared/interceptors';
+import { httpInterceptorProviders } from './core/interceptors/index';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
